+++ NEWS.md
@@ -1,5 +1,9 @@
 # ggtrace (development version)
 
+### **Improvements**
+
+- 
+
 # ggtrace 0.3.1
 
 ### **Improvements**